lib: move DeploymentID to own file
