+++ lib/deployment.go
@@ -3,9 +3,7 @@ package sous
 // has problems? go:generate ggen cmap.CMap(cmap.go) sous.Deployments(deployments.go) CMKey:DeployID Value:*Deployment
 
 import (
-	"crypto/md5"
 	"fmt"
-	"io"
 	"strings"
 
 	"github.com/opentable/sous/util/logging"
@@ -39,31 +37,7 @@ type (
 		// Kind is the kind of software that SourceRepo represents.
 		Kind ManifestKind
 	}
-
-	// A DeploymentID identifies a deployment.
-	DeploymentID struct {
-		ManifestID ManifestID
-		Cluster    string
-	}
 )
-
-// Digest genrates an MD5 sum generated by the combination of  Git repo,
-// project flavor, and the name of the cluster separted by null bytes.
-func (depID *DeploymentID) Digest() []byte {
-	h := md5.New()
-	sep := []byte{0}
-	io.WriteString(h, depID.ManifestID.Source.String())
-	h.Write(sep)
-	io.WriteString(h, depID.ManifestID.Flavor)
-	h.Write(sep)
-	io.WriteString(h, depID.Cluster)
-
-	return h.Sum(nil)
-}
-
-func (depID DeploymentID) String() string {
-	return depID.Cluster + ":" + depID.ManifestID.String()
-}
 
 // Clone returns a deep copy of this deployment.
 func (d Deployment) Clone() *Deployment {
