+++ lib/deploymentid.go
@@ -1,0 +1,30 @@
+package sous
+
+import (
+	"crypto/md5"
+	"io"
+)
+
+// A DeploymentID identifies a deployment.
+type DeploymentID struct {
+	ManifestID ManifestID
+	Cluster    string
+}
+
+// Digest genrates an MD5 sum generated by the combination of  Git repo,
+// project flavor, and the name of the cluster separted by null bytes.
+func (depID *DeploymentID) Digest() []byte {
+	h := md5.New()
+	sep := []byte{0}
+	io.WriteString(h, depID.ManifestID.Source.String())
+	h.Write(sep)
+	io.WriteString(h, depID.ManifestID.Flavor)
+	h.Write(sep)
+	io.WriteString(h, depID.Cluster)
+
+	return h.Sum(nil)
+}
+
+func (depID DeploymentID) String() string {
+	return depID.Cluster + ":" + depID.ManifestID.String()
+}
